[release/8.0] [Blazor] Fix default reconnect display not showing after enhanced navigation (#51173)
# Fix default reconnect display not showing after enhanced navigation

Fixes an issue where the Blazor Server default reconnect display stops showing up during reconnection if an enhanced navigation occurs after the first time it was displayed.

## Description

This issue happened because the reconnect display had its own state tracking whether it added itself to the DOM. However, an enhanced navigation may forcefully remove the reconnect display from the DOM without its knowing.

This PR fixes the issue by changing the existing logic to instead check whether the reconnect display's root element is part of the `document`, and if it isn't, it gets added to the `body`.

Fixes #51083

## Customer Impact

If a customer's app has a user with a flaky connection, it's likely that the reconnect display will stop appearing if the user reconnects multiple times in the same session. This means that the user will not receive any feedback that the app is reconnecting and may attempt to keep interacting with the app without understanding why it's not responding.

It should be noted that this bug only applies to "Blazor Web"-style apps, which are new in .NET 8. Existing Blazor Server scenarios are not affected.

## Regression?

- [ ] Yes
- [X] No

This only affects "Blazor Web"-style apps which are new in .NET 8. The reconnect display always exhibited this bug for Blazor Web apps, so this is not a regression.

## Risk

- [ ] High
- [ ] Medium
- [X] Low

The change is extremely small, isolated, and easy to reason about.

## Verification

- [X] Manual (required)
- [X] Automated

## Packaging changes reviewed?

- [ ] Yes
- [ ] No
- [X] N/A
---------

Co-authored-by: Christoph <topfolas@t-online.de>