[release/8.0] [Blazor] Defer enhanced component updates until reconnection circuit completes (#51242)
# Defer enhanced component updates until reconnection circuit completes

Ensures that enhanced updates to interactive server components get deferred until the circuit is reconnected.

## Description

This PR includes two changes:
1. Takes circuit "connected-ness" into account when determining if enhanced component updates can be applied
2. Waits until any given root component's disposal gets acknowledged by .NET before the browser stops tracking it for SSR updates

Fixes #51082

## Customer Impact

This change impacts Blazor Web scenarios where interactive server components exist on the page.

With this fix, it's more likely that a customer with an unstable connection will be able to continue using the app without refreshing the page. In particular:
1. If a developer creates a custom, less intrusive reconnection dialog that allows the user to keep interacting with the app, a user-induced page navigation during reconnection won't bring the app to an unusable state
2. If the user quickly navigates between pages while the circuit is unstable, there won't be console errors caused by attempting to interactively update a component that was removed from the page

## Regression?

- [ ] Yes
- [X] No

These fixes only affect "Blazor Web" scenarios, which are new in .NET 8.

## Risk

- [ ] High
- [X] Medium
- [X] Low

This is a moderately-sized change. However, we have existing automated tests that verify the scenarios that this code is likely to impact, and we've added new tests to help ensure that these fixes introduce the above stability improvements.

## Verification

- [X] Manual (required)
- [X] Automated

## Packaging changes reviewed?

- [ ] Yes
- [ ] No
- [X] N/A