Fix streaming SSR timing issue (#51333)
# Fix streaming SSR timing issue

On certain rare occasions, streaming SSR could produce a blank page due to a timing issue.

## Description

If a page component:

* Has `@attribute [StreamRendering]`
* Is being visited through enhanced navigtion
* Has an `OnInitializedAsync` whose returned task is not pre-completed but does complete very quickly (e.g., `await Task.Yield()`)

... then there was a race condition. On rare occasions (seems to be about 1 in 20 page loads), the result would be a blank page. This didn't show up in our E2E tests because they always wait for fairly long async tasks so that the browser automation had time to observe the loading states. This PR adds a new E2E test that does recreate the issue and reliably fails - it does many, many page loads and basically always hits the problem at least once without the fix included here.

Fixes #51345

## Customer Impact

Without this fix, a particular combination of features would be unreliable.

## Regression?

- [ ] Yes
- [x] No

[If yes, specify the version the behavior has regressed from]

## Risk

- [ ] High
- [ ] Medium
- [x] Low

The only change is to fix what is definitely a bug. It's hard to see how this would have a negative effect, as no other code is built on top of this. It's not a public API - it's just part of how the Razor Components endpoint works.

## Verification

- [x] Manual (required)
- [x] Automated

## Packaging changes reviewed?

- [ ] Yes
- [ ] No
- [x] N/A